requirements.txt
flask
flask-cors
python-dotenv
google-genai
tqdm
requests
fastapi
uvicorn
python-multipart
